+++ src/components/Footer.tsx
@@ -1,5 +1,11 @@
+import FaIcon from '@/components/FaIcon'
+import { AUTHOR } from '@/defaults'
+
 export default function Footer() {
   return <div className="footer small">
-    <a href="https://github.com/jm-sky/" target="_blank" rel="noopener noreferrer">@jm-Sky</a>
+    <a href={AUTHOR.url} target="_blank" rel="noopener noreferrer">
+      <FaIcon icon="github" mr />
+      {AUTHOR.title}
+    </a>
   </div>
 }
