+++ src/defaults.ts
@@ -1,0 +1,5 @@
+export const BREAK_LENGTH = 5
+export const SESSION_LENGTH = 25
+export const INTERVAL_AMOUNT = 1000
+export const AUDIO_SRC = '/assets/beepSound.wav'
+export const AUTHOR = { title: 'jm-Sky', url: 'https://github.com/jm-sky/' }
