--- src/index.css
@@ -1,11 +1,0 @@
-body {
-  width: 100vw;
-  height: 100vh;
-  margin: 0;
-  display: grid;
-  place-items: center;
-  background-color: #222;
-  background: radial-gradient(circle at 0% 0%, #0000 25%, #eef1 30%, #0000 35%),
-              radial-gradient(circle at 100% 100%, #0000 10%, #eef1 20%, #0000 30%),
-              radial-gradient(circle at bottom, #44444a, #22222a, #22222a);
-}
