+++ src/scss/index.scss
@@ -1,0 +1,50 @@
+:root {
+  --clr-primary: rgb(0, 135, 255);
+  --clr-primary-dark: rgba(10, 20, 35, 0.5);
+  --clr-warning: #e33;
+}
+
+body {
+  font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
+  width: 100vw;
+  height: 100vh;
+  margin: 0;
+  display: grid;
+  place-items: center;
+  background-color: #222;
+  background: radial-gradient(circle at 0% 0%, #0000 25%, #eef1 30%, #0000 35%),
+              radial-gradient(circle at 100% 100%, #0000 10%, #eef1 20%, #0000 30%),
+              radial-gradient(circle at bottom, #44444a, #22222a, #22222a);
+}
+
+.language-switcher {
+  position: absolute;
+  top: 1rem;
+  right: 1rem;
+  background-color: #111a;
+  padding: .5rem 1rem;
+  border: #aaaa;
+  border-radius: 1rem;
+  color: #fff;
+  display: flex;
+  align-items: center;
+  gap: .5rem;
+  box-shadow: #0005 0 .25rem .5rem;
+
+  button {
+    border: 0;
+    background: transparent;
+    color: white;
+    border-radius: .5rem;
+    padding: .2rem .3rem;
+
+    &:hover {
+      opacity: .85;
+    }
+
+    &.active {
+      color: var(--clr-primary);
+      outline: 1px solid var(--clr-primary);
+    }
+  }
+}
