+++ vite.config.ts
@@ -1,7 +1,13 @@
 import { defineConfig } from 'vite'
+import { resolve } from "path"
 import react from '@vitejs/plugin-react'
 
 // https://vitejs.dev/config/
 export default defineConfig({
+  resolve: {
+    alias: {
+      "@": resolve(__dirname, "./src"), 
+    },
+  },
   plugins: [react()],
 })
